"Products", "Assets", "Registration", "ProductD.."
Product details. And some images for the website and some other stuff. 
There is also the page for the registration
Hope it is PHP you are using for the backend i can write some php for the user auth and verification if you dont mind so you can join with the end points and make the required changes sir